fix: address react doctor sponsor ads findings
